Update unraid_array_fan.sh
Fixed while statements using Bing AI
 
For testing, I've set the variables manually in shell, up to line 82 of the script.
Why is there a problem with the \, while the command without the \ does return the proper temperature (32) of the $CURRENT_DRIVE, which I have manually set to 1 in this case?

root@Juiceb0x:/# /usr/sbin/smartctl -A ${HD[$CURRENT_DRIVE]} | grep -m 1 -i Temperature_Celsius | awk '{print $10}'
32
root@Juiceb0x:/# smartctl -A ${HD[$CURRENT_DRIVE]} | grep -m 1 -i Temperature_Celsius | awk '{print \$10}'
awk: cmd. line:1: {print \$10}
awk: cmd. line:1:        ^ backslash not last character on line

also, is there any obvious problem with the following code, which is lines 85 through 100 of the script?

# while loop to get the highest temperature of active drives. 
# If all are spun down then high temp will be set to 0.
while [ "$CURRENT_DRIVE" -le "$NUM_OF_DRIVES" ]
do
  SLEEPING='hdparm -C ${HD[$CURRENT_DRIVE]} | grep -c standby'
  if [ "$SLEEPING" == "0" ]; then
    CURRENT_TEMP="smartctl -A ${HD[$CURRENT_DRIVE]} | grep -m 1 -i Temperature_Celsius | awk '{print \$10}'"
    OUTPUT+=" -- Drive "${HD[$CURRENT_DRIVE]}" temp "$CURRENT_TEMP$'\n'
    if [ "$HIGHEST_TEMP" -le "$CURRENT_TEMP" ]; then
      HIGHEST_TEMP=$CURRENT_TEMP
    fi
  fi
#echo $CURRENT_TEMP
  let "CURRENT_DRIVE+=1"
done
OUTPUT+="Highest temp is: "$HIGHEST_TEMP$'\n'


The issue you're encountering is due to how the shell is interpreting the commands within the variables. When you're assigning a command to a variable as a string (like `SLEEPING` and `CURRENT_TEMP`), the command isn't being executed at that time. Instead, the command is being stored as a plain string.

So, when you're checking `if [ "$SLEEPING" == "0" ]`, it's actually comparing the string `'hdparm -C ${HD[$CURRENT_DRIVE]} | grep -c standby'` to `"0"`, which will always be false.

Similarly, `CURRENT_TEMP` is storing the command as a string, not the output of the command. So when you're appending `CURRENT_TEMP` to `OUTPUT`, it's actually appending the string, not the temperature value.

To fix this, you should use command substitution `$(command)` when assigning the commands to the variables. This will execute the command and assign the output to the variable. Here's the corrected code:

```bash
while [ "$CURRENT_DRIVE" -le "$NUM_OF_DRIVES" ]
do
  SLEEPING=$(hdparm -C ${HD[$CURRENT_DRIVE]} | grep -c standby)
  if [ "$SLEEPING" == "0" ]; then
    CURRENT_TEMP=$(smartctl -A ${HD[$CURRENT_DRIVE]} | grep -m 1 -i Temperature_Celsius | awk '{print $10}')
    OUTPUT+=" -- Drive "${HD[$CURRENT_DRIVE]}" temp "$CURRENT_TEMP$'\n'
    if [ "$HIGHEST_TEMP" -le "$CURRENT_TEMP" ]; then
      HIGHEST_TEMP=$CURRENT_TEMP
    fi
  fi
  let "CURRENT_DRIVE+=1"
done
OUTPUT+="Highest temp is: "$HIGHEST_TEMP$'\n'
```

This should resolve the issue and allow your script to correctly capture and compare the drive temperatures. Let me know if you need further assistance! 😊